third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/AnimationEffectTiming/iterationStart-expected.txt

This is a testharness.js-based test.
PASS Test default value
FAIL Test that changing the iterationStart affects computed timing when backwards-fillingassert_equals: expected 0.5 but got 0.4999999999999999
FAIL Test that changing the iterationStart affects computed timing during the active phaseassert_equals: expected 0.5 but got 0.4999999999999999
FAIL Test that changing the iterationStart affects computed timing when forwards-fillingassert_equals: expected 0.5 but got 0.4999999999999996
PASS Using invalid values should throw TypeError
Harness: the test ran to completion.
